requirements.txt
anyio==3.6.2
autopep8==2.0.0
click==8.1.3
fastapi==0.85.2
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing-extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
